+++ crates/forge/src/scarb/config.rs
@@ -20,8 +20,10 @@ pub struct ForgeConfigFromScarb {
     pub detailed_resources: bool,
     /// Save execution traces of all test which have passed and are not fuzz tests
     pub save_trace_data: bool,
-    /// Builds profile of all test which have passed and are not fuzz tests
+    /// Build profiles of all tests which have passed and are not fuzz tests
     pub build_profile: bool,
+    /// Generate a coverage report for the executed tests which have passed and are not fuzz tests
+    pub coverage: bool,
     /// Fork configuration profiles
     pub fork: Vec<ForkTarget>,
     /// Limit of steps
@@ -67,8 +69,11 @@ pub(crate) struct RawForgeConfig {
     /// Save execution traces of all test which have passed and are not fuzz tests
     pub save_trace_data: bool,
     #[serde(default)]
-    /// Builds profiles of all test which have passed and are not fuzz tests
+    /// Build profiles of all tests which have passed and are not fuzz tests
     pub build_profile: bool,
+    #[serde(default)]
+    /// Generate a coverage report for the executed tests which have passed and are not fuzz tests
+    pub coverage: bool,
     #[serde(default)]
     /// Fork configuration profiles
     pub fork: Vec<RawForkTarget>,
@@ -138,6 +143,7 @@ impl TryFrom<RawForgeConfig> for ForgeConfigFromScarb {
             detailed_resources: value.detailed_resources,
             save_trace_data: value.save_trace_data,
             build_profile: value.build_profile,
+            coverage: value.coverage,
             fork: fork_targets,
             max_n_steps: value.max_n_steps,
         })